+++ bootique-jersey-client-instrumented/src/main/java/io/bootique/jersey/client/instrumented/threshold/JerseyHealthChecks.java
@@ -6,7 +6,7 @@ import io.bootique.metrics.health.HealthCheckGroup;
 import java.util.Map;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class JerseyHealthChecks implements HealthCheckGroup {
     private Map<String, HealthCheck> healthChecks;